refactor: 컴포넌트 폴더 구조 정리 및 common/Header.tsx 삭제
- common/Header.tsx 삭제
- header/ 내 컴포넌트 전체를 layout/으로 이동
- common/AuthProvider.tsx 를 context/로 이동
- CompactHeader.tsx 비활성화(return null)
