+++ src/components/layout/AuthHeader.tsx
