+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router";
 import routes from "./routes.tsx";
 import "./index.css";
-import { AuthProvider } from "@/components/common/AuthProvider.tsx";
+import { AuthProvider } from "@/components/context/AuthProvider.tsx";
 
 const router = createBrowserRouter(routes);
